แก้ไขพวกหน้า nevbar  แต่Userยังลบได้อยู่

+++ editEventTitle.php
@@ -1,21 +1,35 @@
 <?php
 
 require_once('connect.php');
+$mysqli = new mysqli("localhost","root","root","calender");
+if ($mysqli -> connect_errno) {
+  echo "Failed to connect to MySQL: " . $mysqli -> connect_error;
+  exit();
+}
 if (isset($_POST['delete']) && isset($_POST['id'])){
 	
 	$id = $_POST['id'];
 	
-	$sql = "DELETE FROM events WHERE id = $id";
-	$query = $bdd->prepare( $sql );
-	if ($query == false) {
-	 print_r($bdd->errorInfo());
-	 die ('Erreur prepare');
-	}
-	$res = $query->execute();
-	if ($res == false) {
-	 print_r($query->errorInfo());
-	 die ('Erreur execute');
-	}
+	$strSQL = "SELECT * FROM account WHERE userLevel = 'admin' ";
+	$objQuery = mysqli_query($bdd,$strSQL);
+	$objResult = mysqli_fetch_array($objQuery);
+	if($objResult)
+	{
+			$sql = "DELETE FROM events WHERE id = $id";
+			$query = $bdd->prepare( $sql );
+			if ($query == false) {
+		 	print_r($bdd->errorInfo());
+		 	die ('Erreur prepare');
+			}
+			$res = $query->execute();
+			if ($res == false) {
+		 	print_r($query->errorInfo());
+			die ('Erreur execute');
+	}		 
+	else{
+		echo '<script>alert("กรุณาติดต่อผู้ดูแลระบบ")window.location = "calendar.php";</script>';
+		}
+}
 	
 }elseif (isset($_POST['title']) && isset($_POST['color']) && isset($_POST['id'])){
 	
@@ -37,5 +51,5 @@ if (isset($_POST['delete']) && isset($_POST['id'])){
 	}
 
 }
-echo '<script>alert("ลบเรียบร้อยแล้ว");window.location = "index.php";</script>';
-?>
+
+
